+++ frontend/src/metabase/admin/permissions/components/PermissionsTable/PermissionsTable.jsx
@@ -4,7 +4,7 @@ import PropTypes from "prop-types";
 import Tooltip from "metabase/components/Tooltip";
 import Modal from "metabase/components/Modal";
 import ConfirmContent from "metabase/components/ConfirmContent";
-import Ellipsified from "metabase/components/Ellipsified";
+import Ellipsified from "metabase/core/components/Ellipsified";
 
 import { PermissionsSelect } from "../PermissionsSelect";
 import {
