+++ frontend/src/metabase/admin/settings/components/widgets/CustomGeoJSONWidget.jsx
@@ -6,7 +6,7 @@ import { t } from "ttag";
 import Utils from "metabase/lib/utils";
 import Select, { Option } from "metabase/core/components/Select";
 import Confirm from "metabase/components/Confirm";
-import Ellipsified from "metabase/components/Ellipsified";
+import Ellipsified from "metabase/core/components/Ellipsified";
 import Modal from "metabase/components/Modal";
 import LoadingAndErrorWrapper from "metabase/components/LoadingAndErrorWrapper";
