+++ frontend/src/metabase/components/Breadcrumbs.jsx
@@ -5,7 +5,7 @@ import { Link } from "react-router";
 import S from "./Breadcrumbs.css";
 
 import Icon from "metabase/components/Icon";
-import Ellipsified from "metabase/components/Ellipsified";
+import Ellipsified from "metabase/core/components/Ellipsified";
 
 import cx from "classnames";
