+++ frontend/src/metabase/components/EntityItem.jsx
@@ -6,7 +6,7 @@ import cx from "classnames";
 import EntityMenu from "metabase/components/EntityMenu";
 import Swapper from "metabase/components/Swapper";
 import CheckBox from "metabase/core/components/CheckBox";
-import Ellipsified from "metabase/components/Ellipsified";
+import Ellipsified from "metabase/core/components/Ellipsified";
 import Icon from "metabase/components/Icon";
 import Tooltip from "metabase/components/Tooltip";
 import { isItemPinned } from "metabase/collections/utils";
