+++ frontend/src/metabase/core/components/Ellipsified/Ellipsified.styled.tsx
