+++ frontend/src/metabase/core/components/Ellipsified/Ellipsified.tsx
