+++ frontend/src/metabase/core/components/Ellipsified/index.ts
