+++ frontend/src/metabase/home/components/RecentViews.jsx
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 import { t } from "ttag";
 
 import Icon from "metabase/components/Icon";
-import Ellipsified from "metabase/components/Ellipsified";
+import Ellipsified from "metabase/core/components/Ellipsified";
 import Link from "metabase/core/components/Link";
 import SidebarSection from "./SidebarSection";
 import * as Urls from "metabase/lib/urls";
