+++ frontend/src/metabase/home/homepage/components/HomeModelCard/HomeModelCard.styled.tsx
@@ -1,7 +1,7 @@
 import styled from "@emotion/styled";
 import { color } from "metabase/lib/colors";
 import Icon from "metabase/components/Icon";
-import Ellipsified from "metabase/components/Ellipsified";
+import Ellipsified from "metabase/core/components/Ellipsified";
 
 export const CardIcon = styled(Icon)`
   display: block;
