+++ frontend/src/metabase/visualizations/components/TableInteractive.jsx
@@ -35,7 +35,7 @@ import { zoomInRow } from "metabase/query_builder/actions";
 import ExplicitSize from "metabase/components/ExplicitSize";
 import MiniBar from "./MiniBar";
 
-import Ellipsified from "metabase/components/Ellipsified";
+import Ellipsified from "metabase/core/components/Ellipsified";
 import DimensionInfoPopover from "metabase/components/MetadataInfo/DimensionInfoPopover";
 
 const HEADER_HEIGHT = 36;
