+++ frontend/src/metabase/visualizations/components/legend/LegendCaption.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import { iconPropTypes } from "metabase/components/Icon";
 import Tooltip from "metabase/components/Tooltip";
-import Ellipsified from "metabase/components/Ellipsified";
+import Ellipsified from "metabase/core/components/Ellipsified";
 import LegendActions from "./LegendActions";
 import {
   LegendCaptionRoot,
